Adding README.MD

+++ bubble.h
@@ -12,11 +12,11 @@ namespace Bubble
 	}
 
 	template<typename T>
-	const bool bubbleSort(T* arr, const unsigned int length)
+	const bool bubbleSort(T* arr, const size_t length)
 	{
 		bool isSorted = false;
 
-		for (unsigned int i = 0; i < length - 1; i++) // each pass
+		for (size_t i = 0; i < length - 1; i++) // each pass
 		{
 			for (unsigned int j = 0; j < length - 1 - i; j++) // each comparison
 			{
@@ -33,7 +33,7 @@ namespace Bubble
 			}
 		}
 
-		for (unsigned int i = 0; i < length; i++)
+		for (size_t i = 0; i < length; i++)
 		{
 			std::cout << arr[i] << std::endl;
 		}
@@ -41,11 +41,11 @@ namespace Bubble
 	}
 
 	template<typename T>
-	const bool bubbleSortAndFindnthBiggest(T* arr, const unsigned int length, const unsigned int nth) // third param for finding 'specified'th biggest value in arr
+	const bool bubbleSortAndFindnthBiggest(T* arr, const size_t length, const unsigned int nth) // third param for finding 'specified'th biggest value in arr
 	{
 		bool isSorted = false;
 
-		for (unsigned int i = 0; i < nth; i++)
+		for (size_t i = 0; i < nth; i++)
 		{
 			for (unsigned int j = 0; j < length - 1 - i; j++)
 			{
@@ -63,7 +63,7 @@ namespace Bubble
 		}
 
 		std::cout << arr[length - nth] << std::endl << std::endl;
-		for (unsigned int i = 0; i < length; i++)
+		for (size_t i = 0; i < length; i++)
 		{
 			std::cout << arr[i] << std::endl;
 		}
